module8-task1 - правка по замечаниям прошлых заданий
